svelte-migrate: reamed files
