+++ package.json
@@ -36,5 +36,8 @@
 		"tslib": "^2.3.1",
 		"typescript": "~4.6.2"
 	},
-	"type": "module"
+	"type": "module",
+	"dependencies": {
+		"vite": "^3.1.6"
+	}
 }
