+++ src/routes/api/auth/+server.ts
@@ -1,6 +1,7 @@
+import { json as json$1 } from '@sveltejs/kit';
 import { serialize } from "cookie";
 
-export async function post({ request }) {
+export async function POST({ request }) {
   const response = await fetch(`${process.env.VITE_BACKEND_URL}/auth`, {
     method: "POST",
     body: request.body,
@@ -12,17 +13,14 @@ export async function post({ request }) {
   const { status, message, data } = await response.json();
 
   if (status === "error") {
-    return {
-      status: response.status,
-      body: { status, message }
-    }
+    return json$1({ status, message }, {
+      status: response.status
+    })
   }
 
-  return {
-    status: 200,
-    body: {
-      status, message
-    },
+  return json$1({
+  status, message
+}, {
     headers: {
       "Set-Cookie": serialize('session_token', data, {
         path: '/',
@@ -31,5 +29,5 @@ export async function post({ request }) {
         secure: process.env.NODE_ENV === "production",
       })
     }
-  }
+  })
 }
