+++ src/routes/api/logout/+server.ts
@@ -1,17 +1,16 @@
+import { json } from '@sveltejs/kit';
 import { serialize } from "cookie";
 
-export async function post() {
-  return {
-    status: 200,
+export async function POST() {
+  return json({
+  message: "Signed out."
+}, {
     headers: {
       'Set-Cookie': serialize('session_token', null, {
         path: '/',
         maxAge: -1,
         httpOnly: true
       })
-    },
-    body: {
-      message: "Signed out."
     }
-  };
+  });
 }
